Changes from NyX: - Several python scripts are modified to make it work in NyX smoothly

+++ Examples/slurm_run.py
@@ -20,7 +20,7 @@ job_directory = "%s" %os.getcwd()
 mkdir_p(job_directory+'/job')
 
 # list of AOD/nPCA
-tau = -np.logspace(np.log10(0.01), np.log10(2.0), 3)
+tau = -np.logspace(np.log10(0.01), np.log10(2.0), 7)
 # splitting into chunks to make the code efficient and run easily in DISCOVER
 try:
     arrayNum= int(sys.argv[1])
@@ -28,7 +28,7 @@ except:
     print('No array number is given: it should be 0,1 or 2 \n Using default value 0')
     arrayNum = 0
 
-npca_ = [range(0,36), range(36,72), range(72, 107)]
+npca_ = [range(0,22), range(22,44), range(44, 66), range(66, 88), range(88, 107)] # modified for NyX
 npca = npca_[arrayNum] # max is 107
 
 # Solar Zenith angle (used if no real sun-satellite geometry is used)
@@ -38,7 +38,7 @@ sza = list(itertools.chain.from_iterable(itertools.repeat(x, 12) for x in sza_))
 # realGeometry: True if using the real geometry provided in the .mat file
 useRealGeometry = 1
 # Job name
-jobName = 'T%d_' %arrayNum # 'A' for 2modes, 'Z' for realGeometry
+jobName = 'M%d' %arrayNum # 'A' for 2modes, 'Z' for realGeometry
 if not useRealGeometry: jobName = jobName + str(SZA); varStr = 'aod'
 else: varStr = 'nPCA'
 
@@ -57,11 +57,11 @@ for aod in tau:
     mkdir_p(job_directory)
 
     with open(job_file, 'w') as fh:
-        fh.writelines("#!/usr/local/bin/bash\n\n")
+        fh.writelines("#!/usr/bin/bash\n\n")
         fh.writelines("#SBATCH --job-name=%s%.4d\n" % (jobName, aod_))
         fh.writelines("#SBATCH --output=./job/%s_%.4d.out.%s\n" % (jobName, aod_, '%A'))
         fh.writelines("#SBATCH --error=./job/%s_%.4d.err.%s\n" % (jobName, aod_, '%A'))
-        fh.writelines("#SBATCH --time=02:59:59\n")
+        fh.writelines("#SBATCH --time=23:59:59\n")
         # In Discover
         if 'discover' in hostname:
             fh.writelines('#SBATCH --constraint="sky"\n')
@@ -70,6 +70,11 @@ for aod in tau:
         # In Uranus
         elif 'uranus' in hostname:
             fh.writelines("#SBATCH --partition=LocalQ\n")
+        # in NyX
+        elif 'nyx' in hostname:
+            fh.writelines("#SBATCH --partition=zen4\n")
+            fh.writelines("#SBATCH --ntasks=24\n")
+            fh.writelines("#SBATCH --mem=120G\n")
         # fh.writelines("#SBATCH --qos=short\n")
         # fh.writelines("#SBATCH --nodes=1\n")
         # fh.writelines("#SBATCH --ntasks-per-node=1\n\n")
@@ -81,6 +86,10 @@ for aod in tau:
         fh.writelines("date\n")
         if useRealGeometry:
             if 'discover' in hostname:
+                for i in npca:
+                    fh.writelines("python runRetrievalSimulationSlurm.py %d %s %s %s %2.3f &\n" %(int(i), instrument,
+                                                                                           SZA, useRealGeometry, aod))
+            elif 'nyx' in hostname:
                 for i in npca:
                     fh.writelines("python runRetrievalSimulationSlurm.py %d %s %s %s %2.3f &\n" %(int(i), instrument,
                                                                                            SZA, useRealGeometry, aod))
@@ -98,7 +107,8 @@ for aod in tau:
         fh.writelines("wait\n")
         fh.writelines("echo 0\n")
         fh.writelines("echo End: \n")
-        fh.writelines("rm -rf ${TMPDIR}\n")
+        if 'discover' in hostname:
+            fh.writelines("rm -rf ${TMPDIR}\n")
         fh.writelines("date")
 
     fh.close()
@@ -108,4 +118,4 @@ for aod in tau:
         os.system("sbatch %s" %job_file)
     else:
     	print('<><><><> dry run, check the ./job directory for slurm files <><><><>')
-print('Jobs sumitted succesfully check the ./job/ folder for output/error')
+print('Jobs submitted successfully check the ./job/ folder for output/error')
